resolved feedbacks from previous PR
