+++ src/components/Book.jsx
@@ -1,22 +1,18 @@
-// eslint-disable
-
 import React from 'react';
-import Proptypes from 'prop-types';
 import { useDispatch } from 'react-redux';
+import Proptypes from 'prop-types';
 import { removeBook } from '../redux/books/booksSlice';
+import { checkStatus } from '../redux/categories/categoriesSlice';
 
 function Book({ title, author, id }) {
   const dispatch = useDispatch();
-  const handleStatus = () => {
-    console.log('Check status');
-  };
   return (
     <div>
       <h2>{title}</h2>
       <span>{author}</span>
       <div className="actions">
         <button type="button" onClick={() => dispatch(removeBook(id))}>Remove</button>
-        <button type="button" onClick={handleStatus}>Check Status</button>
+        <button type="button" onClick={() => dispatch(checkStatus())}>Check Status</button>
       </div>
     </div>
   );
