+++ src/routes/Books.jsx
@@ -1,27 +1,10 @@
-import React, { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import Book from '../components/Book';
 import Form from '../components/Form';
-import { addBook } from '../redux/books/booksSlice';
-
-const initialValues = {
-  title: '',
-  author: '',
-  category: '',
-};
 
 function Books() {
-  const [state, setState] = useState(initialValues);
   const booksArray = useSelector((state) => state.booksReducer.books);
-  const dispatch = useDispatch();
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setState((prev) => ({ ...prev, [name]: value, item_id: `item${booksArray.length + 1}` }));
-  };
-  const handleSubmit = () => {
-    dispatch(addBook(state));
-  };
   return (
     <>
       {booksArray.map(
@@ -34,7 +17,7 @@ function Books() {
           />
         ),
       )}
-      <Form handleDispatch={handleSubmit} handleChange={handleChange} />
+      <Form />
     </>
   );
 }
